+++ example-crates/tiny/src/main.rs
@@ -2,20 +2,8 @@
 
 #![no_std]
 #![no_main]
-#![allow(internal_features)]
-#![feature(lang_items)]
-#![feature(core_intrinsics)]
 
-extern crate compiler_builtins;
 extern crate origin;
-
-#[panic_handler]
-fn panic(_panic: &core::panic::PanicInfo<'_>) -> ! {
-    core::intrinsics::abort()
-}
-
-#[lang = "eh_personality"]
-extern "C" fn eh_personality() {}
 
 #[no_mangle]
 unsafe fn origin_main(_argc: usize, _argv: *mut *mut u8, _envp: *mut *mut u8) -> i32 {