+++ src/arch/aarch64.rs
@@ -5,6 +5,7 @@ use core::arch::asm;
 #[cfg(relocation_model = "pic")]
 use linux_raw_sys::elf::{Elf_Dyn, Elf_Ehdr};
 #[cfg(feature = "signal")]
+#[cfg(test)]
 use linux_raw_sys::general::__NR_rt_sigreturn;
 #[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
 #[cfg(relocation_model = "pic")]
@@ -16,35 +17,33 @@ use {
     rustix::thread::RawPid,
 };
 
-/// The program entry point.
-///
-/// # Safety
-///
-/// This function must never be called explicitly. It is the first thing
-/// executed in the program, and it assumes that memory is laid out according
-/// to the operating system convention for starting a new program.
 #[cfg(feature = "origin-start")]
-#[naked]
-#[no_mangle]
-pub(super) unsafe extern "C" fn _start() -> ! {
+naked!(
+    "
+    The program entry point.
+
+    # Safety
+
+    This function must never be called explicitly. It is the first thing
+    executed in the program, and it assumes that memory is laid out according
+    to the operating system convention for starting a new program.
+    ";
+    pub(super) fn _start() -> !;
+
     // Jump to `entry`, passing it the initial stack pointer value as an
     // argument, a null return address, a null frame pointer, and an aligned
     // stack pointer. On many architectures, the incoming frame pointer is
     // already null.
-    asm!(
-        "mov x0, sp",   // Pass the incoming `sp` as the arg to `entry`.
-        "mov x30, xzr", // Set the return address to zero.
-        "b {entry}",    // Jump to `entry`.
-        entry = sym super::program::entry,
-        options(noreturn),
-    )
-}
+    "mov x0, sp",   // Pass the incoming `sp` as the arg to `entry`.
+    "mov x30, xzr", // Set the return address to zero.
+    "b {entry}";    // Jump to `entry`.
+    entry = sym super::program::entry;
+    options(noreturn)
+);
 
 /// Abort the process without involving any panic handling code.
 ///
 /// This is a stable equivalent to `core::intrinsics::abort()`.
-#[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
-#[cfg(relocation_model = "pic")]
 pub(super) fn abort() -> ! {
     unsafe {
         asm!("brk #0x1", options(noreturn, nostack));
@@ -258,13 +257,14 @@ pub(super) const TLS_OFFSET: usize = 0;
 #[cfg(feature = "thread")]
 #[inline]
 pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usize) -> ! {
+    assert_eq!(__NR_exit, 93); // TODO: obviate this
     asm!(
         "svc 0",
         "mov x0, xzr",
-        "mov x8, {__NR_exit}",
+        "mov x8, 93", // TODO: use {__NR_exit}
         "svc 0",
         "udf #16",
-        __NR_exit = const __NR_exit,
+        //__NR_exit = const __NR_exit, // TODO: Use this when `asm_const` is stabilized.
         in("x8") __NR_munmap,
         in("x0") map_addr,
         in("x1") map_len,
@@ -272,23 +272,29 @@ pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usiz
     );
 }
 
-/// Invoke the `__NR_rt_sigreturn` system call to return control from a signal
-/// handler.
-///
-/// # Safety
-///
-/// This function must never be called other than by the `sa_restorer`
-/// mechanism.
 #[cfg(feature = "signal")]
-#[naked]
-pub(super) unsafe extern "C" fn return_from_signal_handler() {
-    asm!(
-        "mov x8, {__NR_rt_sigreturn}",
-        "svc 0",
-        "udf #16",
-        __NR_rt_sigreturn = const __NR_rt_sigreturn,
-        options(noreturn)
-    );
+naked!(
+    "
+    Invoke the `__NR_rt_sigreturn` system call to return control from a signal
+    handler.
+
+    # Safety
+
+    This function must never be called other than by the `sa_restorer`
+    mechanism.
+    ";
+    pub(super) fn return_from_signal_handler() -> ();
+
+    "mov x8, 139", // TODO: use {__NR_rt_sigreturn}
+    "svc 0",
+    "udf #16";
+    //__NR_rt_sigreturn = const __NR_rt_sigreturn // TODO: Use this when `asm_const` is stabilized.
+    ;
+    options(noreturn)
+);
+#[cfg(test)] // TODO: obviate this
+fn test_rt_sigreturn() {
+    assert_eq!(__NR_rt_sigreturn, 139);
 }
 
 /// Invoke the appropriate system call to return control from a signal