+++ src/arch/x86.rs
@@ -1,6 +1,12 @@
 //! Architecture-specific assembly code.
 
+#[cfg(feature = "thread")]
+#[cfg(not(feature = "nightly"))]
+use crate::ptr::{without_provenance_mut, Polyfill as _};
 use core::arch::asm;
+#[cfg(feature = "thread")]
+#[cfg(feature = "nightly")]
+use core::ptr::without_provenance_mut;
 #[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
 #[cfg(relocation_model = "pic")]
 use linux_raw_sys::elf::{Elf_Dyn, Elf_Ehdr};
@@ -8,48 +14,46 @@ use linux_raw_sys::elf::{Elf_Dyn, Elf_Ehdr};
 #[cfg(relocation_model = "pic")]
 use linux_raw_sys::general::{__NR_mprotect, PROT_READ};
 #[cfg(feature = "signal")]
+#[cfg(test)]
 use linux_raw_sys::general::{__NR_rt_sigreturn, __NR_sigreturn};
 #[cfg(feature = "thread")]
 use {
     core::ffi::c_void,
-    core::ptr::without_provenance_mut,
     linux_raw_sys::general::{__NR_clone, __NR_exit, __NR_munmap},
     rustix::thread::RawPid,
 };
 
-/// The program entry point.
-///
-/// # Safety
-///
-/// This function must never be called explicitly. It is the first thing
-/// executed in the program, and it assumes that memory is laid out according
-/// to the operating system convention for starting a new program.
 #[cfg(feature = "origin-start")]
-#[naked]
-#[no_mangle]
-pub(super) unsafe extern "C" fn _start() -> ! {
+naked!(
+    "
+    The program entry point.
+
+    # Safety
+
+    This function must never be called explicitly. It is the first thing
+    executed in the program, and it assumes that memory is laid out according
+    to the operating system convention for starting a new program.
+    ";
+    pub(super) fn _start() -> !;
+
     // Jump to `entry`, passing it the initial stack pointer value as an
     // argument, a null return address, a null frame pointer, and an aligned
     // stack pointer. On many architectures, the incoming frame pointer is
     // already null.
-    asm!(
-        "mov eax, esp", // Save the incoming `esp` value.
-        "push ebp",     // Pad for stack pointer alignment.
-        "push ebp",     // Pad for stack pointer alignment.
-        "push ebp",     // Pad for stack pointer alignment.
-        "push eax",     // Pass saved the incoming `esp` as the arg to `entry`.
-        "push ebp",     // Set the return address to zero.
-        "jmp {entry}",  // Jump to `entry`.
-        entry = sym super::program::entry,
-        options(noreturn),
-    )
-}
+    "mov eax, esp", // Save the incoming `esp` value.
+    "push ebp",     // Pad for stack pointer alignment.
+    "push ebp",     // Pad for stack pointer alignment.
+    "push ebp",     // Pad for stack pointer alignment.
+    "push eax",     // Pass saved the incoming `esp` as the arg to `entry`.
+    "push ebp",     // Set the return address to zero.
+    "jmp {entry}";  // Jump to `entry`.
+    entry = sym super::program::entry;
+    options(noreturn)
+);
 
 /// Abort the process without involving any panic handling code.
 ///
 /// This is a stable equivalent to `core::intrinsics::abort()`.
-#[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
-#[cfg(relocation_model = "pic")]
 pub(super) fn abort() -> ! {
     unsafe {
         asm!("ud2", options(noreturn, nostack));
@@ -355,15 +359,16 @@ pub(super) const TLS_OFFSET: usize = 0;
 #[cfg(feature = "thread")]
 #[inline]
 pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usize) -> ! {
+    assert_eq!(__NR_exit, 1); // TODO: obviate this
     asm!(
         // Use `int 0x80` instead of vsyscall, since vsyscall would attempt to
         // touch the stack after we `munmap` it.
         "int 0x80",
         "xor ebx, ebx",
-        "mov eax, {__NR_exit}",
+        "mov eax, 1", // TODO: use {__NR_exit}
         "int 0x80",
         "ud2",
-        __NR_exit = const __NR_exit,
+        //__NR_exit = const __NR_exit, // TODO: Use this when `asm_const` is stabilized.
         in("eax") __NR_munmap,
         in("ebx") map_addr,
         in("ecx") map_len,
@@ -371,23 +376,30 @@ pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usiz
     );
 }
 
-/// Invoke the `__NR_rt_sigreturn` system call to return control from a signal
-/// handler.
-///
-/// # Safety
-///
-/// This function must never be called other than by the `sa_restorer`
-/// mechanism.
 #[cfg(feature = "signal")]
-#[naked]
-pub(super) unsafe extern "C" fn return_from_signal_handler() {
-    asm!(
-        "mov eax, {__NR_rt_sigreturn}",
-        "int 0x80",
-        "ud2",
-        __NR_rt_sigreturn = const __NR_rt_sigreturn,
-        options(noreturn)
-    );
+naked!(
+    "
+    Invoke the `__NR_rt_sigreturn` system call to return control from a signal
+    handler.
+
+    # Safety
+
+    This function must never be called other than by the `sa_restorer`
+    mechanism.
+    ";
+    pub(super) fn return_from_signal_handler() -> ();
+
+    "mov eax, 173", // TODO: use {__NR_rt_sigreturn}
+    "int 0x80",
+    "ud2";
+    //__NR_rt_sigreturn = const __NR_rt_sigreturn // TODO: Use this when `asm_const` is stabilized.
+    ;
+    options(noreturn)
+);
+#[cfg(feature = "signal")]
+#[cfg(test)] // TODO: obviate this
+fn test_rt_sigreturn() {
+    assert_eq!(__NR_rt_sigreturn, 173);
 }
 
 /// Invoke the appropriate system call to return control from a signal
@@ -398,14 +410,28 @@ pub(super) unsafe extern "C" fn return_from_signal_handler() {
 /// This function must never be called other than by the `sa_restorer`
 /// mechanism.
 #[cfg(feature = "signal")]
-#[naked]
-pub(super) unsafe extern "C" fn return_from_signal_handler_noinfo() {
-    asm!(
-        "pop eax",
-        "mov eax, {__NR_sigreturn}",
-        "int 0x80",
-        "ud2",
-        __NR_sigreturn = const __NR_sigreturn,
-        options(noreturn)
-    );
+naked!(
+    "
+    Invoke the `__NR_sigreturn` system call to return control from a signal
+    handler.
+
+    # Safety
+
+    This function must never be called other than by the `sa_restorer`
+    mechanism.
+    ";
+    pub(super) fn return_from_signal_handler() -> ();
+
+    "pop eax",
+    "mov eax, 119", // TODO: use {__NR_sigreturn}
+    "int 0x80",
+    "ud2";
+    //__NR_sigreturn = const __NR_sigreturn // TODO: Use this when `asm_const` is stabilized.
+    ;
+    options(noreturn)
+);
+#[cfg(feature = "signal")]
+#[cfg(test)] // TODO: obviate this
+fn test_sigreturn() {
+    assert_eq!(__NR_sigreturn, 119);
 }