+++ test-crates/origin-start/src/bin/tls.rs
@@ -3,30 +3,16 @@
 
 #![no_std]
 #![no_main]
-#![allow(internal_features)]
-#![feature(lang_items)]
-#![feature(core_intrinsics)]
 #![feature(thread_local)]
 #![feature(strict_provenance)]
 
 extern crate alloc;
-extern crate compiler_builtins;
 
 use alloc::boxed::Box;
-use atomic_dbg::dbg;
 use core::cell::UnsafeCell;
 use core::arch::asm;
 use core::ptr::{addr_of_mut, without_provenance_mut};
 use origin::{program, thread};
-
-#[panic_handler]
-fn panic(panic: &core::panic::PanicInfo<'_>) -> ! {
-    dbg!(panic);
-    core::intrinsics::abort();
-}
-
-#[lang = "eh_personality"]
-extern "C" fn eh_personality() {}
 
 #[global_allocator]
 static GLOBAL_ALLOCATOR: rustix_dlmalloc::GlobalDlmalloc = rustix_dlmalloc::GlobalDlmalloc;